requirements.txt
Django==3.2.8
djangorestframework==3.12.4
graphene-django==2.15.0